// MCP_Frontend-main/src/app/shared/models/deposit-product.ts
export class DepositProduct {
    id: number;
    prodCode: string;
    description: string;
    currencyId: number;
    minimumAge: number;
    maximumAge: number;
    acctNoSchemeId: number;
    gender: string;
    allowMultipleAccount: string;
    keepDailyAccruedInterest: string;
    interestBearing: string;
    bearingMethod: string;
    interestBearingFrequencyValue: number;
    interestBearingFrequencyCode: string;
    maximumAccountPerProduct: number;
    minimumOpeningDeposit: number;
    minimumDepositAmount: number;
    minimumWithdrawalAmount: number;
    accountCycleValue: number;
    accountCycleCode: string;
    chargeCycleValue: number;
    chargeCycleCode: string;
    riskId: number;
    minimumAccountBalance: number;
    allowOverdraw: string;
    dormancyPeriodCode: string;
    dormancyPeriodValue: number;
    freeWithdrawalAllowed: number;
    withdrawPeriod: string;
    withdrawValue: number;
    maturityPeriodCode: string;
    maturityPeriodValue: number;
    maturityInstruction: string;
    createdBy: string;
    status: string;
    modifyDate: Date;
    createDate: Date;
    uuid: string;
    modifiedBy: string;
    edit: Boolean;
}